requirements.txt
azure-ai-formrecognizer>=4.0.0
python-dotenv~=1.0.0
requests~=2.31.0